Add some Unit Tests and fix `SensorsListener`
This commit will add unit test for  `SensorsConsumer` and `SensorsListener` classes.
In oreder to better test SensorsListener the class was changed to extract the internal HandlerThread, that now the user must create, and it was renamed from `SensorsListenerThread` to `SensorsListener`.
In this commit was also fixed a bug indise `SensorsListener`'s method `startListening()` that caused an error if one of the rensors was not available in the device.§