Change `SensorsData` to be more easily testable
This commit will make some changes to the `SensorsData` class to make it easily testable. The main change is using the "Humble Object" pattern to separate all the internal windowing logic from the thread-safeness; this is done creating an internal class called `SensorsDataImpl` that takes care of the window data and can be tested normally and letting `SensorsData` being responsible of synchronizing the acces from all threads.

Another change reguards the accelerometer/gyroscope sensors samples passed to the method, befor it was a simple array of unspecifyed length, now it uses the `SensorSample` class that has x,y,z parameters and can be constructed from an array using `fromArray()` method.