Extract sensor's constants out to `SensorService`
This commit will extract all sensors related constants previously inside `SensorsData` and `SensorsListener` out of this classes and in the father `SensorService`. This is done to make them easily customizable having all the constants in one place, but without using a global object containing all the constants; this is not customizable if we extract the seonsors code in it's own library.