Add new module for `recorder` wear app
This commit will add a new module for the Data Recorder wear app.
This app is different from the standard `wear` one, and has the
purpose to help us record a new dataset.

This app will be used only by trusted user, and it's more a side
tool than a real published application.