Recorder: Add `RecorderViewModel` and `RecorderRepository`
This commit will add a ViewModel and a Repository classes that help manage
the state of the UI and the data production.

In particular the `RecorderViewModel` is responsible of keeping track of the
UI states. For example if we are recoding the UI should display the recording
screen, othervise the main actions list should be displayed.

The `RecorderRepository` will produce the data, with this we mean that this class
is responsible of calling the recorder service that records the particular activity
and stored the data in a database.