Sensorslib: Rename `SensorsListener` to `SensorEventProducer`
This commit will rename the class `SensorsListener` to `SensorEventProducer` and refactor every use
of it throuhout the code.

This commit will also chahge the constructor of `SensorEventProducer` to take a sampling period in ms
and the sampling window size in order to compute the max latency in ms multiplying the two togheter. This
way we have one less configuration parameter to pass around and we reduce the risk of mistakes.