Sensorlib: Rename `SensorsListener` to `SensorEventProducer`
This commit will rename `SensorsListener` to `SensorEventProducer` and refactor
it's use throuhout the code.

Also in this commit the constructor of `SensorEventProducer` is changed to take a
sampling period in milliseconds and a sampling window size in order to compute the
max latency multiplying the two togheter. This way we have one less parameter to
pass around and we reduce the risk of mistakes.