Reorder build dependencies
This commit will fix and reorder all gradle build dependencies for the root project and the submodules.

In the root `build.gradle` we have declared a variable for each dependency version used in the submodules
so we can easily manage all the common version in one place. In the root module we have also declared junit5
and ktlint plugins to be applyed by the submodule.

In general this commit will reorder all the imports to be more glancable.