Sensorlib: Remove old `SensorXXX` files in favor of new API.
This commit will remove all the files related to the old API (`SensorReaderHelper`,
`SensorDataHandler`, ...) in favor of the new API (`SensorWindowProducer`, ...).

`SensorWindowProducer` is changed in order to accomodate setting his `OnNewWindow` listener
outside the constructor. This is particulary usefull in the method `asFlow` that will be
included in the newxt commit.