Sensorlib: Fix bug in `SensorWindowBuffer` producing half windows
This commit will fix a bug in `SensorWindowBuffer` acusing it to signal that
a window was ready when in reality it was only half-done.

This was caused by a check in `pushAccelerometer` and `pushGyroscope` that was
checking only if one of `accIndex` or `gyroIndex` was exceeding before returning
the window. In case we were receiving some accelerometer samples, and then all of
hte gyroscope ones (ore vice-versa) this will fire not waiting for all the accelerometer
samples and resulting in a corrupted window with some zeros as values.

In this commit we are also adding a parameter to `SensorSampleFilter` to denote the
name of the sensor being filtered in debug logs.